feat(discovery): modularize etcd discovery
